+++ lib/smarter_csv.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'set'
+
 require "smarter_csv/version"
 require "smarter_csv/file_io"
 require "smarter_csv/options_processing"