+++ spec/features/hash_transformations/remove_not_mapped_keys_spec.rb
@@ -5,31 +5,70 @@ require 'spec_helper'
 fixture_path = 'spec/fixtures'
 
 describe ':remove_unmapped_keys option' do
-  it 'it has no effect on loading a file without options' do
-    options = {}
-    data = SmarterCSV.process("#{fixture_path}/lots_of_columns.csv", options)
-    expect(data.size).to eq 1
-    expect(data.first.size).to eq 474 # there are some empty rows in the fixture
+  describe 'v1.x behavior' do
+    let(:options) { { } }
+
+    it 'it has no effect on loading a file without options' do
+      options = {}
+      data = SmarterCSV.process("#{fixture_path}/lots_of_columns.csv", options)
+      expect(data.size).to eq 1
+      expect(data.first.size).to eq 474 # there are some empty rows in the fixture
+    end
+
+    it 'it has no effect if provided without :key_mapping' do
+      options[:remove_unmapped_keys] = true
+      data = SmarterCSV.process("#{fixture_path}/lots_of_columns.csv", options)
+      expect(data.size).to eq 1
+      expect(data.first.size).to eq 474 # there are some empty rows in the fixture
+    end
+
+    it 'it defaults to false and has no effect if :key_mapping is provided without :remove_unmapped_keys' do
+      options[:key_mapping] = {column_0: :one, column_15: :two, column_42: :three}
+      data = SmarterCSV.process("#{fixture_path}/lots_of_columns.csv", options)
+      expect(data.size).to eq 1
+      expect(data.first.size).to eq 474 # there are some empty rows in the fixture
+    end
+
+    it 'it removes non-mapped keys/columns when set to true and :key_mapping is provided' do
+      options[:remove_unmapped_keys] = true
+      options[:key_mapping] = {column_0: :one, column_15: :two, column_42: :three}
+      data = SmarterCSV.process("#{fixture_path}/lots_of_columns.csv", options)
+      expect(data.size).to eq 1
+      expect(data.first.size).to eq 2 # column_15 is empty
+    end
   end
 
-  it 'it has no effect if provided without :key_mapping' do
-    options = {remove_unmapped_keys: true}
-    data = SmarterCSV.process("#{fixture_path}/lots_of_columns.csv", options)
-    expect(data.size).to eq 1
-    expect(data.first.size).to eq 474 # there are some empty rows in the fixture
-  end
+  describe 'v2.x behavior' do
+    let(:options) { {v2_mode: true } }
+    let(:mapping) { {column_0: :one, column_15: :two, column_42: :three} }
 
-  it 'it defaults to false and has no effect if :key_mapping is provided without :remove_unmapped_keys' do
-    options = {key_mapping: {column_0: :one, column_15: :two, column_42: :three}}
-    data = SmarterCSV.process("#{fixture_path}/lots_of_columns.csv", options)
-    expect(data.size).to eq 1
-    expect(data.first.size).to eq 474 # there are some empty rows in the fixture
-  end
+    it 'it has no effect on loading a file without options' do
+      data = SmarterCSV.process("#{fixture_path}/lots_of_columns.csv", options)
+      expect(data.size).to eq 1
+      expect(data.first.size).to eq 474 # there are some empty rows in the fixture
+    end
 
-  it 'it removes non-mapped keys/columns when set to true and :key_mapping is provided' do
-    options = {remove_unmapped_keys: true, key_mapping: {column_0: :one, column_15: :two, column_42: :three}}
-    data = SmarterCSV.process("#{fixture_path}/lots_of_columns.csv", options)
-    expect(data.size).to eq 1
-    expect(data.first.size).to eq 2 # column_15 is empty
+    it 'it has no effect if provided without :key_mapping' do
+      options[:remove_unmapped_keys] = true
+      data = SmarterCSV.process("#{fixture_path}/lots_of_columns.csv", options)
+      expect(data.size).to eq 1
+      expect(data.first.size).to eq 474 # there are some empty rows in the fixture
+    end
+
+    it 'it defaults to false and has no effect if :key_mapping is provided without :remove_unmapped_keys' do
+      options[:header_transformations] = [:keys_as_symbols, {key_mapping: mapping}]
+
+      data = SmarterCSV.process("#{fixture_path}/lots_of_columns.csv", options)
+      expect(data.size).to eq 1
+      expect(data.first.size).to eq 474 # there are some empty rows in the fixture
+    end
+
+    it 'it removes non-mapped keys/columns when set to true and :key_mapping is provided' do
+      options[:remove_unmapped_keys] = true
+      options[:header_transformations] = [:keys_as_symbols, {key_mapping: mapping}]
+      data = SmarterCSV.process("#{fixture_path}/lots_of_columns.csv", options)
+      expect(data.size).to eq 1
+      expect(data.first.size).to eq 2 # column_15 is empty
+    end
   end
 end