+++ spec/features/header_handling/silence_missing_keys_spec.rb
@@ -10,7 +10,7 @@ describe 'silence missing keys' do
   subject(:read_csv) { SmarterCSV.process("#{fixture_path}/silence_missing_keys.csv", options) }
 
   it 'prints warning message for missing keys by default' do
-    expect(SmarterCSV).not_to receive(:puts)
+    expect(SmarterCSV).to receive(:puts).with(/DEPRECATED OPTION.*key_mapping/)
     expect{ read_csv }.to raise_exception(
       SmarterCSV::KeyMappingError, "ERROR: can not map headers: THIS, missing_key"
     )
@@ -18,7 +18,7 @@ describe 'silence missing keys' do
 
   it 'maps the keys from the CSV file correctly' do
     options[:silence_missing_keys] = true
-    expect(SmarterCSV).not_to receive(:puts)
+    expect(SmarterCSV).to receive(:puts).with(/DEPRECATED OPTION.*key_mapping/)
     data = SmarterCSV.process("#{fixture_path}/silence_missing_keys.csv", options)
     expect(data.size).to eq 1
     expect(data[0].keys).to eq %i[this that]