+++ spec/smarter_csv/file_encoding_spec.rb
@@ -61,4 +61,3 @@ RSpec.describe SmarterCSV do
     end
   end
 end
-